Refactor TCP server as other transports have with base class and read implementation
